refactor fetch test metadata to util function
