+++ src/api/evergreen.js
@@ -1,6 +1,7 @@
 // @flow strict
 
 import type { EvergreenLog, EvergreenTaskLogType } from "../models";
+import { getTestMetadata } from "../util";
 import { EVERGREEN_BASE } from "../config";
 import { stringToInteralEvergreenTaskLogType } from "../models";
 
@@ -59,41 +60,16 @@ export function taskURL(taskID: string, execution: ?number): string {
   return base + `/${execution}`;
 }
 
-const getTestMetadataURL = ({
-  taskId,
-  execution,
-  testId,
-}: {
-  taskId: string,
-  execution: number,
-  testId: string,
-}) =>
-  `${EVERGREEN_BASE}/rest/v2/tasks/${taskId}/tests?execution=${execution}&test_id=${testId}`;
-
 export async function fetchEvergreen(log: EvergreenLog): Promise<Response> {
   const init = { method: "GET", credentials: "include" };
   let req = "";
   if (log.type === "evergreen-task") {
     req = new Request(taskLogRawURL(log.id, log.execution, log.log), init);
   } else if (log.type === "evergreen-test") {
-    const { taskId, testId, execution } = log;
-    const testMetadataUrl = getTestMetadataURL({ execution, taskId, testId });
-    const testMetadataReq = new Request(testMetadataUrl);
-    const res = await window.fetch(testMetadataReq, {
-      credentials: "include",
-    });
-    const data = await res.json();
-    let testLogUrl = "";
-    if (Array.isArray(data) && data.length) {
-      if (data.length > 1) {
-        console.error(
-          "Multiple results error: multiple results came back from the rest v2 tests endpoint instead of 1. Url:",
-          testMetadataUrl
-        );
-      }
-      testLogUrl = data[0].logs.url_raw_display;
-      req = new Request(`${EVERGREEN_BASE}${testLogUrl}`, init);
-    }
+    const { execution, testId, taskId } = log;
+    const data = await getTestMetadata({ execution, testId, taskId });
+    const testLogUrl = data.logs.url_raw_display;
+    req = new Request(`${EVERGREEN_BASE}${testLogUrl}`, init);
   } else if (log.type === "evergreen-test-by-name") {
     req = new Request(
       testLogByNameRawURL(log.task, log.execution, log.test),
